Integration of Jukebox Library (#451) (#457)
* add Jukebox library

todo:
- saving / loading of new format
- inferrence of unknown data like past versions
- move the temporary float casts to proper use of double
- make sound related functions take double for timing
- inform people that the Jukebox sound player was renamed to SoundByte lol

* make sound, input scheduling, and super curve use double precision

* successfully load charts

* editor works again

v1 riqs can be saved and loaded

* first tempo and volume markers are unmovable

fix loading of charts' easing values

* use gsync / freesync

* update Jukebox refs to SoundByte

* game events use double part 1

Air Rally - Glee Club converted

* don't load song if chart load fails

* finish conversion of all minigames

* remove editor waveform toggle

* timeline now respects added song offset length

clear cache files on app close
prepped notes for dsp sync

* update timeline length when offset changed

* update to latest Jukebox

* make error panel object in global game manager

* improve conductor music scheduling

* added error message box

fix first game events sometimes not playing

+++ Assets/Scripts/Games/Kitties/Kitties.cs
@@ -127,7 +127,7 @@ namespace HeavenStudio.Games
 
         }
 
-        public void Clap(bool isMice, bool isInverse, bool keepSpawned, float beat, int type)
+        public void Clap(bool isMice, bool isInverse, bool keepSpawned, double beat, int type)
         {
             player.ScheduleClap(beat, type);
             MultiSound.Play(new MultiSound.Sound[] {
@@ -215,7 +215,7 @@ namespace HeavenStudio.Games
             }
         }
 
-        public void Roll(bool keepSpawned, float beat)
+        public void Roll(bool keepSpawned, double beat)
         {
             if (!player.canClap)
                 return;
@@ -265,7 +265,7 @@ namespace HeavenStudio.Games
             }
         }
 
-        public void CatchFish(float beat)
+        public void CatchFish(double beat)
         {
             //if (!player.canClap)
             //    return;
@@ -438,7 +438,7 @@ namespace HeavenStudio.Games
                 player.canClap = false;
         }
 
-        public void InstantSpawn(bool isMice, bool isInverse, float beat, int pos)
+        public void InstantSpawn(bool isMice, bool isInverse, double beat, int pos)
         {
             BeatAction.New(Cats[0], new List<BeatAction.Action>()
                 {
